+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Next release
 
+### Added
+
+- Projects on homepage can now be filtered by 'Active', 'Archived' or 'All' through a tab system.
 
 ## 2022-02-01